loadtest/requirements.txt
locustio==0.13.0